vagrant: parameterize for native build as well as Vagrant build
Notes on strategy:

git clone is only used for Vagrant builds, hence the {%if's in states that
require it.  git_clone.sls is still present, but gutted for non-vagrant
(we assume Jenkins or the build system has checked out the right branch
to vars.builddir.

gitname exists to allow the Vagrant and non-Vagrant builds to use
different names for the build git repo, since in the case of Vagrant
it's a cloned copy, whereas non-Vagrant builds in the externally-cloned
checked-out version.  The base of the repo is vars.builddir/vars.gitname.

pkgdest is where the output ends up (/git, the shared dir for Vagrant,
or "parent of builddir/gitname", i.e. builddir, for native) and where
the build-output tar is copied in from.

(cherry picked from commit b303336f9f7a0cf90da651472ca34c95f4c5eb02)